backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.0.3
neo4j==5.14.1
pymongo==4.6.0
pypdf==3.17.1
python-docx==1.1.0
spacy==3.7.2
openai-whisper==20231117
pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2
# Audio processing
librosa==0.10.1
pydub==0.25.1
# OCR for images
pytesseract==0.3.10
# Supabase client
supabase==2.3.4
# Neo4j driver for knowledge graph
neo4j==5.14.1
# Additional ML dependencies for analytics
scikit-learn==1.3.2
# WebSocket support is included in FastAPI
# Additional async HTTP client
httpx==0.25.2
# For real-time integrations
websockets==12.0
# UUID support is built-in to Python